Move the questionnaire activity to the separate application data capture gallery
