+++ example/src/main/AndroidManifest.xml
@@ -26,17 +26,16 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".QuestionnaireActivity">
+        <activity
+                android:name=".PatientListActivity"
+                android:label="@string/title_patient_list"
+                android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".PatientListActivity"
-            android:label="@string/title_patient_list"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity
             android:name=".PatientDetailActivity"
             android:label="@string/title_patient_detail"
